refactor: trends table
