+++ src/components/SocialGrid/Item.js
@@ -8,7 +8,7 @@ import Button from '@santiment-network/ui/Button'
 import { createTrigger } from '../../ducks/Signals/common/actions'
 import { buildInTrendingWordsSignal } from '../../ducks/Signals/utils/utils'
 import LoginPopup from '../banners/feature/PopupBanner'
-import WordCloud from '../WordCloud/WordCloud'
+import WordCloud from '../WordCloud/WordCloudWithHeader'
 import DarkTooltip from '../Tooltip/DarkTooltip'
 import { dividePhraseInWords } from './topics'
 import Chart from './Chart'
